requirements.txt
alembic==1.10.2
apispec==6.3.0
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib-inline==0.1.6
packaging==23.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
prompt-toolkit==3.0.38
psycopg2==2.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.1
requests==2.28.2
rq==1.13.0
six==1.16.0
SQLAlchemy==2.0.5.post1
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
webargs==8.2.0
Werkzeug==2.2.3
zipp==3.15.0
